Safely calling base class functions on instances, added `ExposableClass`
Previously, calling a base class function on a class with multiple inheritance was likely to cause a segmentation fault due to incorrect memory addresses. This was due to the raw usage of `reinterpret_cast` on a `SQUserPointer` (`void*`).

The `ExposableClass` class was added to ensure safely calling exposed functions on other classes a class inherits from. To be able to have that safety, `ExposableClass` needs to be inherited by every exposable class as a whole.

Now, a `SQUserPointer` (`void*`) is first `reinterpret_cast`-ed to a pointer to the base class `ExposableClass`, which we are sure the instance class inherits from, since it's required. Afterwards, the resulting `ExposableClass*` is either `static_cast`-ed or `dynamic_cast`-ed, based on whether the type required for the function derives from `ExposableClass` (or not, which would indicate it's an unrelated class, derived from in multiple inheritance, which is where `dynamic_cast` is used).