Fix native class member variables overriding root table
The root table is now added as a delegate to both the "_get" and "_set" tables of a class, created by simplesquirrel, when adding a member variable to it.

To get an entry from the "_get" or "_set" table, first it's searched for without employing delegation. If found, that means it's a member variable, exposed from C++.
If nothing is found, a second search is done, this time employing delegation. If found, the entry originates from the root table, added as a delegate, meaning it should only be pushed to the stack.

Fixes #2.