Always initialize "_get" and "_set" tables for classes
Since commit 67479ab makes these tables use the root table as a delegate, they are now initialized by default, removing the requirement for a member variable to be added for that to happen.

Initializing these tables by default allows for consistent behaviour of classes, in relation to delegation, with and without exposed member variables, since delegation will now always be present.