Fix error converting Squirrel array to `std::vector`
Because of the `convert()` and `convertRaw()` functions of `Array` not popping each value, popped from the array, from the stack, the function would fail if the array has more than 1 item. Additionally, those functions also used to push values to the C++ `std::vector` in reverse order.

A new approach has been taken, where the functions will no longer clone an array and pop from it to iterate it, but rather would perform an iteration using `sq_next()` (like it's done with tables).